OpResolverService: Add first preliminary version of OpResolverService
An OpResolver will try to resolve all dependencies of an Op or an pool
of Ops. Resolving means, that he will create a directed graph of ops with
a single source. If the source is updated, the ops in the graph will
be recalculated.

TODO: I'm not sure if I like the API, yet.